Reorganised Test Cases into Directory, added more
