+++ stylometryproject/stylometryapp/tests/test_verify.py
@@ -1,0 +1,1 @@
+# TO DO, IMPLEMENT VERIFICATION TESTS (awaiting finalisation of algorithm output)
