+++ stylometryproject/stylometryapp/views.py
@@ -49,7 +49,6 @@ def create_profile(request):
 
     if request.method == 'POST':
         new_profile_name = request.POST.get('name')
-        print("Creating profile with name: " + new_profile_name)
 
         # Create a new profile
         profile = Profile(name=new_profile_name, user=request.user)
